Inflation injection only handles classes, not types
You cannot express anything for which a type would be needed in XML. Only raw classes. The factory should still support parameterized types, though, which it currently does not. A failing test case is included as ignored.

Additionally, nested types are specified using their reflection names (since they're looked up with Class.forName) and so the key needs to use the reflection name.