Avoid wrapping a factory in a provider
In the case of our own @AssistedInject.Factory or Dagger's @AssistedFactory, when used as a dependency request in another factory do not wrap it in a Provider. The factory is already stateless, so wrapping it in a stateless provider is pointless and wasteful. Moreover, Dagger's assisted injection implementation explicitly forbids this with an error.